+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/rtc/PeerChannel.kt
@@ -37,6 +37,9 @@ import java.util.concurrent.Executors
 import java.util.zip.DeflaterInputStream
 
 interface PeerChannel {
+    // SoraMediaChannel からアクセスするためのプロパティ
+    val localVideoManager: RTCLocalVideoManager?
+
     fun handleInitialRemoteOffer(
         offer: String,
         mid: Map<String, String>?,
@@ -178,7 +181,7 @@ class PeerChannelImpl(
 
     private val sdpConstraints = componentFactory.createSDPConstraints()
     private val localAudioManager = componentFactory.createAudioManager()
-    private val localVideoManager = componentFactory.createVideoManager()
+    override val localVideoManager = componentFactory.createVideoManager(appContext)
 
     private var videoSender: RtpSender? = null
     private var audioSender: RtpSender? = null
@@ -201,6 +204,13 @@ class PeerChannelImpl(
     // setRemoteDescription() を実行すると encodings が変更前に戻ってしまう
     // そのため re-offer, update 時に再度 encodings をセットする
     private var offerEncodings: List<Encoding>? = null
+
+    /**
+     * resumeAudioRecording() 実行時に、PeerConnection の
+     * setAudioRecording(true) を実行する必要があるかどうかを示す変数。
+     * この変数は Sora 接続時に音声のハードミュートが有効な場合に利用する。
+     */
+    private var needsPeerConnectionAudioRecordingStart: Boolean = mediaOption.audioOption.initialAudioHardMute
 
     init {
         val compressedDataChannels =
@@ -532,6 +542,10 @@ class PeerChannelImpl(
     }
 
     override fun requestClientOfferSdp(): Single<Result<SessionDescription>> {
+        // NOTE: setup() は 2 箇所 (requestClientOfferSdp(), handleInitialRemoteOffer()) から呼ばれるため、
+        // onAddLocalStream() が二重に呼ばれる可能性があるように見えるが、実際には
+        // SoraMediaChannel では requestClientOfferSdp を呼ぶ場合に listener を持たない PeerChannelImpl を
+        // 使っているため、 このメソッド呼び出し経由で setupInternal() が呼ばれても listener.onAddLocalStream() は呼ばれない。
         return setup().flatMap {
             SoraLogger.d(TAG, "requestClientOfferSdp")
             return@flatMap createClientOfferSdp()
@@ -609,12 +623,20 @@ class PeerChannelImpl(
         localAudioManager.track?.let {
             localStream.addTrack(it)
         }
+        if (mediaOption.audioOption.initialAudioHardMute && mediaOption.audioOption.audioDeviceModule == null) {
+            // Sora 接続時に音声のハードミュートを有効化するフラグが立っている場合、ミュート状態にする
+            audioRecordingPaused = true
+            // 録音を開始しないために setAudioRecording(false) を呼ぶ
+            conn?.setAudioRecording(false)
+        }
 
         SoraLogger.d(TAG, "local managers' initTrack: video => ${mediaOption.videoUpstreamContext}")
         localVideoManager?.initTrack(factory!!, mediaOption.videoUpstreamContext, appContext)
         localVideoManager?.track?.let {
             localStream.addTrack(it)
         }
+        // CameraVideoCapturer を SDK 内で生成している場合はここでキャプチャを開始する
+        localVideoManager?.startOwnedCapture()
 
         SoraLogger.d(TAG, "localStream.audioTracks.size = ${localStream.audioTracks.size}")
         SoraLogger.d(TAG, "localStream.videoTracks.size = ${localStream.videoTracks.size}")
@@ -892,6 +914,13 @@ class PeerChannelImpl(
     }
 
     private suspend fun resumeAudioRecording(): Boolean {
+        // Sora 接続時の音声ハードミュートが有効である場合は PeerConnection の録音を開始する
+        if (needsPeerConnectionAudioRecordingStart) {
+            // false に更新する、この変数はこれ以降更新されない
+            needsPeerConnectionAudioRecordingStart = false
+            // 録音を開始する
+            conn?.setAudioRecording(true)
+        }
         // ADM の録音を再開
         SoraLogger.d(TAG, "[audio_recording_pause] resume ADM audio recording")
         // 既に録音中の場合でも true が返る